Test fitting with daily data
Made data changes:
* Upgraded resolution of synthetic data from weekly to daily, and
* End season on the last Sunday of January to shorten up the dataset
  some.
Interesting result is that `MAge` appears to converge on the `MAge` for
the 2022/23 season regardless of season, which might help to explain
some of the fitting issues. This can be seen in the first trace plot
right after fitting that I added in this notebook in particular. I don't
have a great understanding of how to customize `arviz` plots quickly,
but it should be apparent with the help of the comments on the lines.