Split `docs` job out of `ci` job
Split the `docs` job out of the `ci` job and only run with python 3.12.
The documentation should be independent of version so it serves no value
to build/check it multiple times with different python versions. Also
allows for an early exit if the CI jobs encounter an issue. While the
docs target was removed from the ci target in the make file, the all
target still contains it for rapid developer feedback.