Allow season ranges to be broader than obs
Fix bug where model construction would fail if the observations did not
span the explicit season date range. This bug was introduced when
switching to inferring season start/end date from the min/max date in
the season range. This produces a season range that is a subset of the
explicit ranges. Resolves #79.