Add `propose_p0` to `Curve` class
Added `Curve.propose_p0` as an optional method that subclasses can
implement to propose an initial set of parameters for a single time
series. Also added `get_p0_proposal` as a wrapper method around this
optional method with checks on both the input and outputs. This opens
the door to start working on an initial stage of least squares fitting
by proposing an initial set of parameters that can be used for said
fitting. Heuristics proposed by this method do not need to be "good",
but just close enough.

In addition also made the following minor adjustments:
* Allow the `**kwargs` to `prevalence`, `prevalence_difference` and
  `incidence` methods to accept a float,
* Implemented `propose_p0` method for `LogisticCurve`,
* Removed the deprecated `Curve.evaluate` method.