Move only incidence data supported error
Moved the `NotImplementedError` for if observation types other than
'incidence' found from `_validate_and_format_observations` to
`SeasonalUptakeModel.__init__`. This limitation only has to do with the
seasonal uptake model, not the package as a whole so it makes sense to
move the `NotImplementedError` to where the limitation actually is.
Minor changes made to the test suite as well to account for said change.