Add initial `VaxfluxModel` to switch to `numpyro`
Added an initial skeleton for the `VaxfluxModel` class that is based on
`numpyro` and `jax` rather than `pymc`. Also added new `vaxflux.Curve`
and `vaxflux.LogisticCurve` classes that represent the same curves as
`vaxflux.curves.*` but doing the calculations with `jax`. Began the
`VaxfluxModel` class with an `add_seasons` method to allow users to
specify season ranges as well as a largely placeholder `sample` method.
Other edits include:

- Added `jax` and `numpyro` as dependencies.
- Updated the `scripts/api-reference.py` helper script to generate a
  page for exports directly from `vaxflux`.
- Minor modifications to `mypy` settings to avoid typing `numpyro` and
  be better at typing `numpy`.
- Dropped python 3.13 support as `jax` does not support it. Deferred
  further work on this to #103.
- Added a few entries to the spelling list related to `jax`/`numpyro`.
- Added a `NumericalArrayLike` type alias that is comperable to
  `jax.typing.ArrayLike` but without the boolean types.