Add `VaxfluxModel.add_dates`
Added `add_dates` method for specifying date ranges to a `VaxfluxModel`.
Attempted to use the same underlying infrastructure that `add_seasons`
uses, so work consolidating to private helpers in `vaxflux.dates` is
included. Also added unit tests that are consolidated and performed the
same consolidation for the `add_seasons` unit tests.

Closes #105.